Take all rates from same source
This also extracts the storage logic to be able to test locally
