+++ logger.js
@@ -1,0 +1,10 @@
+const pino = require("pino");
+const logger = pino({
+  prettyPrint: {
+    translateTime: true,
+    colorize: true,
+    ignore: "pid,hostname",
+  },
+});
+
+module.exports = logger
